+++ tests/config/test_config_cli.py
@@ -2,9 +2,9 @@ import time
 from multiprocessing import Process
 
 import pytest
-from click.testing import CliRunner
+from typer.testing import CliRunner
 
-from panoptes.utils.config.cli import config_server_cli
+from panoptes.utils.cli.main import app
 
 
 @pytest.fixture(scope="module")
@@ -21,9 +21,10 @@ def cli_config_port():
 def test_cli_server(runner, config_path, cli_config_port):
     def run_cli():
         result = runner.invoke(
-            config_server_cli,
+            app,
             [
                 "--verbose",
+                "config",
                 "run",
                 "--config-file",
                 f"{config_path}",
@@ -43,9 +44,10 @@ def test_cli_server(runner, config_path, cli_config_port):
     time.sleep(5)
 
     result = runner.invoke(
-        config_server_cli,
+        app,
         [
             "--verbose",
+            "config",
             "--port",
             f"{cli_config_port}",
             "get",
@@ -53,7 +55,6 @@ def test_cli_server(runner, config_path, cli_config_port):
         ],
     )
     assert result.exit_code == 0
-    # Ugh. I hate this. Logger is interfering in annoying ways.
     assert result.stdout.endswith("Testing PANOPTES Unit\n")
 
     proc.terminate()
@@ -62,17 +63,16 @@ def test_cli_server(runner, config_path, cli_config_port):
 
 @pytest.mark.skip("Not working")
 def test_config_server_cli(runner, cli_server, cli_config_port):
-    result = runner.invoke(config_server_cli, ["--verbose", "--port", f"{cli_config_port}", "get", "name"])
+    result = runner.invoke(app, ["--verbose", "config", "--port", f"{cli_config_port}", "get", "name"])
     assert result.exit_code == 0
-    # Ugh. I hate this. Logger is interfering in annoying ways.
     assert result.stdout.endswith("Testing PANOPTES Unit\n")
 
     # Set the name.
-    result = runner.invoke(config_server_cli, ["--port", f"{cli_config_port}", "set", "name", "foobar"])
+    result = runner.invoke(app, ["config", "--port", f"{cli_config_port}", "set", "name", "foobar"])
     assert result.exit_code == 0
     assert result.stdout.endswith("\n{'name': 'foobar'}\n")
 
     # Get the name.
-    result = runner.invoke(config_server_cli, ["--port", f"{cli_config_port}", "get", "name"])
+    result = runner.invoke(app, ["config", "--port", f"{cli_config_port}", "get", "name"])
     assert result.exit_code == 0
     assert result.stdout.endswith("\nfoobar\n")